requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
crayons==0.4.0
cryptography==45.0.5
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jose==1.0.0
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
piexif==1.1.3
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
stegano==2.0.0
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
